fix: There is similar code in two places
